P1.3 static library strings fallback extraction
