+++ mobsf/StaticAnalyzer/views/ios/binary_analysis.py
@@ -15,7 +15,7 @@ from mobsf.StaticAnalyzer.views.ios.classdump import (
 from mobsf.StaticAnalyzer.views.common.binary.lib_analysis import (
     MachOChecksec,
 )
-from mobsf.StaticAnalyzer.views.ios.strings import (
+from mobsf.StaticAnalyzer.views.common.binary.strings import (
     strings_on_binary,
 )
 from mobsf.StaticAnalyzer.views.ios.binary_rule_matcher import (
